Add search service slide animation
Add search service slide animation
