+++ koin/src/main/AndroidManifest.xml
@@ -258,7 +258,7 @@
             android:screenOrientation="portrait"
             android:theme="@style/Theme.AppCompat.Translucent" />
         <activity
-            android:name=".ui.search.SearchActivity"
+            android:name=".ui.search.SearchFragment"
             android:label="코인 커뮤니티"
             android:screenOrientation="portrait"
             android:theme="@style/KAPTheme.NoActionBar" />
