--- koin/src/main/java/in/koreatech/koin/ui/search/SearchActivity.java
@@ -1,161 +1,0 @@
-package in.koreatech.koin.ui.search;
-
-import android.os.Bundle;
-import android.view.KeyEvent;
-import android.view.View;
-import android.view.inputmethod.EditorInfo;
-import android.view.inputmethod.InputMethodManager;
-import android.widget.FrameLayout;
-import android.widget.TextView;
-
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-import androidx.fragment.app.FragmentTransaction;
-
-import butterknife.BindView;
-import butterknife.ButterKnife;
-import butterknife.OnClick;
-import in.koreatech.koin.ui.navigation.KoinNavigationDrawerActivity;
-import in.koreatech.koin.R;
-
-import in.koreatech.koin.core.appbar.AppBarBase;
-import in.koreatech.koin.core.appbar.AppBarSearchBase;
-import in.koreatech.koin.data.network.entity.SearchedArticle;
-import in.koreatech.koin.data.network.interactor.SearchArticleRestInteractor;
-import in.koreatech.koin.ui.search.presenter.SearchArticleContract;
-import in.koreatech.koin.ui.search.presenter.SearchArticlePresenter;
-
-public class SearchActivity extends KoinNavigationDrawerActivity implements AppBarSearchBase.SearchTextChange, SearchArticleContract.View, AppBarSearchBase.SearchEditorAction {
-
-    @BindView(R.id.koin_base_appbar_dark)
-    AppBarSearchBase appbarSearchBase;
-    @BindView(R.id.search_main_framelayout)
-    FrameLayout searchMainFrameLayout;
-
-
-    private FragmentManager fragmentManager;
-    private FragmentTransaction fragmentTransaction;
-    private SearchArticlePresenter articleSearchPresenter;
-    private String currentText;
-
-
-    @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.search_activity_main);
-        ButterKnife.bind(this);
-        init();
-    }
-
-    private void init() {
-        new SearchArticlePresenter(this, new SearchArticleRestInteractor());
-        appbarSearchBase.setOnTextChange(this);
-        appbarSearchBase.setSearchEditorAction(this);
-        showRecentSearch();
-        hideSoftwareKeyBoard();
-    }
-
-    @Override
-    protected void onStart() {
-        super.onStart();
-    }
-
-    @OnClick(R.id.koin_base_appbar_dark)
-    public void koinBaseAppbarClick(View view) {
-        int id = view.getId();
-        if (id == AppBarBase.getLeftButtonId()) {
-            onBackPressed();
-        } else if (id == AppBarBase.getRightButtonId()) {
-            searchCurrentText();
-        }
-    }
-
-    private void changeFragment(Fragment fragment) {
-        if (fragment != null) {
-            fragmentManager = getSupportFragmentManager();
-            fragmentTransaction = fragmentManager.beginTransaction();
-            fragmentTransaction.replace(R.id.search_main_framelayout, fragment).commit();
-        }
-    }
-
-    @Override
-    public void getString(String text) {
-        currentText = text;
-        if (currentText.isEmpty())
-            showRecentSearch();
-    }
-
-    @Override
-    public void showLoading() {
-        showProgressDialog(R.string.loading);
-    }
-
-    @Override
-    public void hideLoading() {
-        hideProgressDialog();
-    }
-
-    @Override
-    public void showEmptyItem() {
-        Fragment fragment = new SearchNoResultFragment();
-        changeFragment(fragment);
-    }
-
-    @Override
-    public void showSearchedArticle(SearchedArticle searchedArticle) {
-        Fragment fragment = new SearchResultFragment();
-        Bundle item = new Bundle();
-        item.putSerializable("SEARCH_ITEMS", searchedArticle);
-        item.putString("CURRENT_TEXT", currentText);
-        fragment.setArguments(item);
-        changeFragment(fragment);
-    }
-
-    @Override
-    public void showRecentSearch() {
-        Fragment fragment = new SearchRecentFragment();
-        changeFragment(fragment);
-    }
-
-    @Override
-    public void showMessage(String message) {
-
-    }
-
-    @Override
-    public void setPresenter(SearchArticlePresenter presenter) {
-        this.articleSearchPresenter = presenter;
-    }
-
-    @Override
-    public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
-        switch (actionId) {
-            case EditorInfo.IME_ACTION_SEARCH:
-                searchCurrentText();
-                break;
-        }
-        return false;
-    }
-
-    public void searchCurrentText() {
-        if (articleSearchPresenter != null && currentText != null) {
-            articleSearchPresenter.getArticleSearched(currentText, 1);
-            articleSearchPresenter.saveText(currentText);
-            hideSoftwareKeyBoard();
-        }
-    }
-
-    public void hideSoftwareKeyBoard() {
-        View view = this.getCurrentFocus();
-        if (view != null) {
-            InputMethodManager inputMethodManager = (InputMethodManager) getSystemService(getApplicationContext().INPUT_METHOD_SERVICE);
-            inputMethodManager.hideSoftInputFromWindow(view.getWindowToken(), 0);
-        }
-    }
-
-    public void setSearchText(String text) {
-        if (text != null)
-            appbarSearchBase.setText(text);
-    }
-}
