+++ koin/src/main/res/layout/search_fragment_main.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.drawerlayout.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/drawer_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/white"
-    android:fitsSystemWindows="true">
+    android:background="@color/white">
 
 
     <LinearLayout
@@ -25,10 +23,8 @@
         <FrameLayout
             android:id="@+id/search_main_framelayout"
             android:layout_width="match_parent"
-            android:layout_height="match_parent" />
+            android:layout_height="match_parent"
+            android:layout_marginBottom="56dp" />
 
-        <include layout="@layout/base_navigation_bar_bottom" />
     </LinearLayout>
-
-    <include layout="@layout/base_navigation_drawer_left" />
-</androidx.drawerlayout.widget.DrawerLayout>
+</FrameLayout>
